requirements.txt
Pillow>=5.2.0
pyusb>=1.0.2
